+++ xclim/indices/_threshold.py
@@ -3,6 +3,7 @@ import xarray
 
 from xclim import run_length as rl
 from xclim import utils
+from xclim.utils import _rolling
 from xclim.utils import declare_units
 from xclim.utils import units
 
@@ -370,7 +371,7 @@ def growing_season_length(
     # compute growth season length on resampled data
     thresh = utils.convert_units_to(thresh, tas)
 
-    c = ((tas > thresh) * 1).rolling(time=window).sum().chunk(tas.chunks)
+    c = _rolling(((tas > thresh) * 1), window=window, dim="time", mode="sum")
 
     def compute_gsl(c):
         nt = c.time.size