+++ HISTORY.rst
@@ -10,6 +10,8 @@ History
 * Enhancement to utils.Indicator to have more dynamic attributes using callables.
 * Added indices `heat_wave_total_length` and `tas` / `tg`.
 * Fixed a bug with typed call signatures that caused downstream failures on library import.
+* Added a `_rolling` util function to fix memory issues on large dask datasets.
+* Added the `subset_shape` function to subset utilities for clipping region-masked datasets via polygons.
 
 0.11.x-beta (2019-10-17)
 ------------------------